Fix. Continuously scrolling background now aligns with player movement.
Previously, we had tried to implement a scrolling background by explicitly shifting the coordinates of the drawn background image every frame.

However, since we're already making use of a panning camera that follows the players movement across the larger game scene, we can simply draw a series of static repeating background images across the scene and allow the camera to take care of the panning internally.

I've currently implemented this very naively (i.e., draw every background image in each frame -- irrespective of whether or not it is visible by the camera)... should be okay for now, but fix if this becomes a performance issue.