Feature. Created a "Stacked Breathing" Bonus.
To reward players for performing a round of stacked breathing (i.e., collecting all of the bubbles in a stacked breathing group), I've created a bonus animation that is fired when the final collectable from any group is collected. I've somewhat arbitrarily chosen 1000 points as the reward for this bonus.