Fix. Remove reference to global variable `nextColor` in BubbleWand.
Previously, the BubbleWand classes relied on a reference to the global variable `nextColor` found in the main application script. This was problematic above and beyond portability because it meant that a shared reference to the color object was stored throughout the script - changing characteristics of the color in one place unintentionally influenced others (e.g., opacity).

BubbleWand now accepts a color as a constructor argument, which can be updated through the `c` property.

TODO: for now, I'm relying on creating a new color object to avoid shared references and unintentional color updates (e.g., bubbleWand.c = color(nextColor.toString());) -- should I actually handle this in the BubbleWand class? Not sure what I think for now.