Improvement. Bubble Area appears as a Paint Can Icon.
Previously, the bottom region of the canvas represented the "paint can" that was used to absorb paint into the wand, encouraging players to perform a deep inhalation before each pep exhalation. However, this was simply a placeholder.

I've improved this feature by adding a "paint can" in the bottom left of the screen, depicted as a colored circle and paint can icon. When players hover over the can and inhale, paint will be absorbed into the wand.

Note, that I've added a black border around the paint wand to ensure that it isn't hidden from view when in mousing over the paint can.